requirements.txt
llama_index==0.6.16.post1
langchain==0.0.188
python-dotenv==1.0.0
faiss-cpu==1.7.4
docx2txt==0.8
pypdf==3.9.0
